[REFACTOR] Supermarket service with interface
